refactor: 마감일이 현재이면 검증 오류 나던걸 현재까지 가능하게 수정
