+++ gunsilver/appcenter-todo-api/src/main/java/com/siillvergun/todolist/global/exception/CustomError.java
@@ -1,4 +1,4 @@
-package com.siillvergun.todolist.global.error;
+package com.siillvergun.todolist.global.exception;
 
 import lombok.Getter;
