+++ gunsilver/appcenter-todo-api/src/main/java/com/siillvergun/todolist/todo/dto/TodoRequestDto.java
@@ -2,7 +2,7 @@ package com.siillvergun.todolist.todo.dto;
 
 import com.siillvergun.todolist.todo.entity.Category;
 import com.siillvergun.todolist.todo.entity.Todo;
-import jakarta.validation.constraints.Future;
+import jakarta.validation.constraints.FutureOrPresent;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
@@ -17,9 +17,9 @@ public class TodoRequestDto {
     @NotBlank(message = "내용은 필수 입니다")
     private String content;
     @NotNull
-    @Future(message = "마감일은 현재 이후여야 합니다")
+    @FutureOrPresent(message = "마감일은 과거이면 안됍니다.")
     private LocalDate dueDate;
-    
+
     @NotNull(message = "정해진 카테고리만 사용할 수 있습니다.")
     private Category category;
